pyproject.toml
[tool.poetry]
name = "organic products"
version = "0.1.0"
description = "Platform where one can get real organic products"
authors = ["LBwaks <obwakuvictor@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.6"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
django-autoslug = "^1.9.9"
django-ckeditor = "^6.7.0"
pillow = "^10.1.0"
django-filter = "^23.3"
crispy-bootstrap5 = "^2023.10"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.2.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
